feature: universal metrics
